Minor codebase structure refactoring
